問題3の解答例

+++ Main.java
@@ -4,16 +4,16 @@ public class Main {
 
     public static void main(String[] args) {
         // 変数宣言・初期化
-        int width, height;
+        long width, height;
         Scanner sc = new Scanner(System.in);
 
         // キーボードからの入力（width/横の長さ）
         System.out.print("width: ");
-        width = sc.nextInt();
+        width = sc.nextLong();
 
         // キーボードからの入力（height/縦の長さ）
         System.out.print("height: ");
-        height = sc.nextInt();
+        height = sc.nextLong();
 
         // 周囲の長さの計算・表示
         System.out.print("circumference: ");
